# Fix race condition in ListenerList.resize(int)